backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-celery-beat==2.6.0
gunicorn==22.0.0
celery==5.4.0
redis==5.0.4
Pillow==10.3.0
anthropic==0.25.0
slack-sdk==3.27.1
slack-bolt==1.18.1
requests==2.32.2
whitenoise==6.6.0
drf-spectacular==0.27.2
flask==3.0.3
flask==3.0.3
